bgk

moment_kinetics.bgk — Module

moment_kinetics.bgk.allowed_wave_amplitude! — Method

moment_kinetics.bgk.init_bgk_pdf! — Method

moment_kinetics.bgk.passing_pdf! — Method

inputs

pdf is the particle distribution function, with the passing part of phase space not filled in

phi_max is the maximum value that e * phi / Te takes

tau = Ti/Te is the ion-electron temperature ratio

x = mvpa^2/2Te + ephi/Te is 1D array containing the total parallel energy (conserved)

output

pdf = particle distribution function; this function fills in the part of phase space where x > e*phi_max/T

moment_kinetics.bgk.setup_dummy_integrals — Method

moment_kinetics.bgk.total_energy_grid — Method

inputs:

vpa = parallel velocity normalized by vts = sqrt(2*Te/ms)

phi = electrostatic potential normalized by Te/e

output: x = vpa^2 + phi is the total parallel energy

moment_kinetics.bgk.trapped_passing_boundary — Method

moment_kinetics.bgk.trapped_pdf! — Method

inputs

phi_max is the maximum value that e * phi / Te takes

tau = Ti/Te is the ion-electron temperature ratio

x = vpa^2 + e*phi is a 2D array containing the total parallel energy on the (z,vpa) grid

y = dummy coordinate for the necessary integrals in the function

integrand = dummy array used to hold integrands defined and integrated in this function

wgts = integration weights associated with y integrals

output

pdf is the particle distribution function for all of phase space, with this function filling in only the part with x < e*phi_max/T