chebyshev

moment_kinetics.chebyshev — Module

moment_kinetics.calculus.elementwise_derivative! — Method

elementwise_derivative!(coord, ff, adv_fac, spectral::chebyshev_info)

Chebyshev transform f to get Chebyshev spectral coefficients and use them to calculate f'.

Note: Chebyshev derivative does not make use of upwinding information within each element.

moment_kinetics.calculus.elementwise_derivative! — Method

elementwise_derivative!(coord, ff, chebyshev::chebyshev_info)

Chebyshev transform f to get Chebyshev spectral coefficients and use them to calculate f'.

moment_kinetics.chebyshev.chebyshev_backward_transform! — Method

moment_kinetics.chebyshev.chebyshev_derivative_single_element! — Method

moment_kinetics.chebyshev.chebyshev_forward_transform! — Method

takes the real function ff on a Chebyshev grid in z (domain [-1, 1]), which corresponds to the domain [π, 2π] in variable theta = ArcCos(z). interested in functions of form f(z) = sumn cn Tn(z) using Tn(cos(theta)) = cos(ntheta) and z = cos(theta) gives f(z) = sumn cn cos(ntheta) thus a Chebyshev transform is equivalent to a discrete cosine transform doing this directly turns out to be slower than extending the domain from [0, 2pi] and using the fact that f(z) must be even (as cosines are all even) on this extended domain, can do a standard complex-to-complex fft fext is an array used to store f(theta) on the extended grid theta ∈ [0,2π) ff is f(theta) on the grid [π,2π] the Chebyshev coefficients of ff are calculated and stored in chebyf n is the number of grid points on the Chebyshev-Gauss-Lobatto grid transform is the plan for the complex-to-complex, in-place fft

moment_kinetics.chebyshev.chebyshev_interpolate_single_element! — Method

moment_kinetics.chebyshev.chebyshev_spectral_derivative! — Method

use Chebyshev basis to compute the first derivative of f

moment_kinetics.chebyshev.chebyshevmoments — Method

compute and return modified Chebyshev moments of the first kind: ∫dx Tᵢ(x) over range [-1,1]

moment_kinetics.chebyshev.chebyshevpoints — Method

returns the Chebyshev-Gauss-Lobatto grid points on an n point grid

moment_kinetics.chebyshev.clenshaw_curtis_weights — Method

returns wgts array containing the integration weights associated with all grid points for Clenshaw-Curtis quadrature

moment_kinetics.chebyshev.scaled_chebyshev_grid — Method

initialize chebyshev grid scaled to interval [-boxlength/2, boxlength/2]

moment_kinetics.chebyshev.setup_chebyshev_pseudospectral — Method

create arrays needed for explicit Chebyshev pseudospectral treatment and create the plans for the forward and backward fast Fourier transforms

moment_kinetics.chebyshev.update_df_chebyshev! — Method

compute the Chebyshev spectral coefficients of the spatial derivative of f

moment_kinetics.chebyshev.update_fcheby! — Method

Chebyshev transform f to get Chebyshev spectral coefficients

moment_kinetics.interpolation.interpolate_to_grid_1d! — Method

Interpolation from a regular grid to a 1d grid with arbitrary spacing

Arguments

result : Array{mkfloat, 1} Array to be overwritten with the result of the interpolation newgrid : Array{mkfloat, 1} Grid of points to interpolate coord to f : Array{mkfloat} Field to be interpolated coord : coordinate coordinate struct giving the coordinate along which f varies chebyshev : chebyshev_info struct containing information for Chebyshev transforms