communication

moment_kinetics.communication — Module

Communication functions and setup

Split the grid into 'blocks'. Each block can use shared memory (MPI shared memory arrays). At the moment only works with a single 'block' containing the whole grid - eventually add more MPI communication functions to communicate between blocks. A block should probably be a 'NUMA region' for maximum efficiency.

Note: charge-exchange collisions loop over neutral species for each ion species. At the moment this loop is not parallelised (although it could be, by introducing some more loop ranges), as at the moment we only run with 1 ion species and 1 neutral species.

moment_kinetics.communication.block_rank — Constant

moment_kinetics.communication.block_size — Constant

moment_kinetics.communication.comm_block — Constant

Communicator connecting a shared-memory region

Must use a Ref{MPI.Comm} to allow a non-const MPI.Comm to be stored. Need to actually assign to this and not just copy a pointer into the .val member because otherwise the MPI.Comm object created by MPI.Comm_split() would be deleted, which probably makes MPI.jl delete the communicator.

moment_kinetics.communication.comm_inter_block — Constant

Communicator connecting the root processes of each shared memory block

Must use a Ref{MPI.Comm} to allow a non-const MPI.Comm to be stored. Need to actually assign to this and not just copy a pointer into the .val member because otherwise the MPI.Comm object created by MPI.Comm_split() would be deleted, which probably makes MPI.jl delete the communicator.

moment_kinetics.communication.comm_world — Constant

Can use a const MPI.Comm for comm_world and just copy the pointer from MPI.COMM_WORLD because MPI.COMM_WORLD is never deleted, so pointer stays valid.

moment_kinetics.communication.global_Win_store — Constant

moment_kinetics.communication.global_rank — Constant

moment_kinetics.communication.global_size — Constant

moment_kinetics.communication.iblock_index — Constant

moment_kinetics.communication.n_blocks — Constant

moment_kinetics.communication.MPISharedArray — Type

moment_kinetics.communication.__init__ — Method

moment_kinetics.communication._block_synchronize — Method

Call an MPI Barrier for all processors in a block.

Used to synchronise processors that are working on the same shared-memory array(s) between operations, to avoid race conditions. Should be (much) cheaper than a global MPI Barrier because it only requires communication within a single node.

Note: some debugging code currently assumes that if blocksynchronize() is called on one block, it is called simultaneously on all blocks. It seems likely that this will always be true, but if it ever changes (i.e. different blocks doing totally different work), the debugging routines need to be updated.

moment_kinetics.communication.allocate_shared — Method

Get a shared-memory array of mk_float (shared by all processes in a 'block')

Create a shared-memory array using MPI.Win_allocate_shared(). Pointer to the memory allocated is wrapped in a Julia array. Memory is not managed by the Julia array though. A reference to the MPI.Win needs to be freed - this is done by saving the MPI.Win into a Vector in the Communication module, which has all its entries freed by the finalize_comms!() function, which should be called when moment_kinetics is done running a simulation/test.

Arguments

dims - mkint or Tuple{mkint} Dimensions of the array to be created. Dimensions passed define the size of the array which is being handled by the 'block' (rather than the global array, or a subset for a single process).

Returns

Array{mk_float}

moment_kinetics.communication.finalize_comms! — Method

Clean up from communications

Do any needed clean-up for MPI, etc. Does not call MPI.Finalize() - this is called anyway when Julia exits, and we do not want to call it explicitly so that multiple runs can be done in a single Julia session.

Frees any shared-memory arrays.

moment_kinetics.communication.free_shared_arrays — Method

moment_kinetics.communication.initialize_comms! — Method

Set up communications

Check that global variables are in the correct state (i.e. caches were emptied correctly if they were used before).

Also does some set up for debugging routines, if they are active.

moment_kinetics.communication.setup_distributed_memory_MPI — Method

Function to take information from user about r z grids and number of processes allocated to set up communicators notation definitions: - block: group of processes that share data with shared memory - z group: group of processes that need to communicate data for z derivatives - r group: group of processes that need to communicate data for r derivatives