file_io

moment_kinetics.file_io — Module

moment_kinetics.file_io.debug_output_counter — Constant

Global counter for calls to debug_dump

moment_kinetics.file_io.debug_output_file — Constant

An nc_info instance that may be initialised for writing debug output

This is a non-const module variable, so does cause type instability, but it is only used for debugging (from debug_dump()) so performance is not critical.

moment_kinetics.file_io.ios — Type

structure containing the various input/output streams

moment_kinetics.file_io.netcdf_info — Type

structure containing the data/metadata needed for netcdf file i/o

moment_kinetics.file_io.debug_dump — Function

debug_dump(ff, dens, upar, ppar, phi, t; istage=0, label="")
debug_dump(fvec::scratch_pdf, fields::em_fields_struct, t; istage=0, label="")

Dump variables into a NetCDF file for debugging

Intended to be called more frequently than write_data_to_binary(), possibly several times within a timestep, so includes a label argument to identify the call site.

Writes to a file called debug_output.cdf in the current directory.

Can either be called directly with the arrays to be dumped (fist signature), or using scratch_pdf and em_fields_struct structs.

nothing can be passed to any of the positional arguments (if they are unavailable at a certain point in the code, or just not interesting). t=nothing will set t to the value saved in the previous call (or 0.0 on the first call). Passing nothing to the other arguments will set that array to 0.0 for this call (need to write some value so all the arrays have the same length, with an entry for each call to debug_dump()).

moment_kinetics.file_io.define_dimensions! — Function

define_dimensions!(fid, nvpa, nvperp, nz, nr, n_species, n_ion_species=nothing,
                   n_neutral_species=nothing)

Define dimensions for an output file.

moment_kinetics.file_io.define_dynamic_variables! — Method

define_dynamic_variables!(fid)

Define dynamic (i.e. time-evolving) variables for an output file.

moment_kinetics.file_io.define_static_variables! — Method

define_static_variables!(vpa,vperp,z,r,composition,collisions,evolve_ppar)

Define static (i.e. time-independent) variables for an output file.

moment_kinetics.file_io.finish_file_io — Method

close all opened output files

moment_kinetics.file_io.input_option_error — Method

accepts an option name which has been identified as problematic and returns an appropriate error message

moment_kinetics.file_io.open_output_file — Method

opens an output file with the requested prefix and extension and returns the corresponding io stream (identifier)

moment_kinetics.file_io.setup_file_io — Method

open the necessary output files

moment_kinetics.file_io.setup_netcdf_io — Method

setup file i/o for netcdf

moment_kinetics.file_io.write_data_to_ascii — Method

moment_kinetics.file_io.write_data_to_binary — Method

write time-dependent data to the netcdf file

moment_kinetics.file_io.write_f_ascii — Method

write the function f(z,vpa) at this time slice

moment_kinetics.file_io.write_fields_ascii — Method

write electrostatic potential at this time slice

moment_kinetics.file_io.write_moments_ascii — Method

write moments of the distribution function f(z,vpa) at this time slice