load_data

moment_kinetics.load_data — Module

moment_kinetics.load_data.load_block_data — Method

moment_kinetics.load_data.load_charged_particle_moments_data — Method

moment_kinetics.load_data.load_coordinate_data — Method

Load data for a coordinate

moment_kinetics.load_data.load_distributed_charged_pdf_slice — Method

Read a slice of an ion distribution function

run_names is a tuple. If it has more than one entry, this means that there are multiple restarts (which are sequential in time), so concatenate the data from each entry together.

The slice to take is specified by the keyword arguments.

moment_kinetics.load_data.load_distributed_neutral_pdf_slice — Method

Read a slice of a neutral distribution function

run_names is a tuple. If it has more than one entry, this means that there are multiple restarts (which are sequential in time), so concatenate the data from each entry together.

The slice to take is specified by the keyword arguments.

moment_kinetics.load_data.load_fields_data — Method

moment_kinetics.load_data.load_input — Method

Load saved input settings

moment_kinetics.load_data.load_mk_options — Method

moment_kinetics.load_data.load_neutral_pdf_data — Method

moment_kinetics.load_data.load_pdf_data — Method

moment_kinetics.load_data.load_rank_data — Method

moment_kinetics.load_data.load_slice — Method

Load a slice of a single variable from a file

moment_kinetics.load_data.load_species_data — Method

moment_kinetics.load_data.load_time_data — Method

If a tuple is given for fid, concatenate the "time" output from each file in the tuple

moment_kinetics.load_data.load_variable — Method

Load a single variable from a file

moment_kinetics.load_data.open_readonly_output_file — Method

moment_kinetics.load_data.read_Dict_from_section — Method

read_Dict_from_section(file_or_group, section_name; ignore_subsections=false)

Read information from section_name in file_or_group, returning a Dict.

By default, any subsections are included as nested Dicts. If ignore_subsections=true they are ignored.

moment_kinetics.load_data.reload_evolving_fields! — Method

Reload pdf and moments from an existing output file.