looping

moment_kinetics.looping — Module

Provides convenience macros for shared-memory-parallel loops

moment_kinetics.looping.loop_ranges — Constant

module variable that we can access by giving fully-qualified name in loop macros

moment_kinetics.looping.loop_ranges_store — Constant

module variable used to store LoopRanges that are swapped into the loopranges variable in begin*_region() functions

moment_kinetics.looping.LoopRanges — Type

LoopRanges structs contain information on which points should be included on this process in loops over shared-memory arrays.

Members

paralleldims::Tuple{Vararg{Symbol}} Indicates which dimensions are (or might be) parallelized when using this LoopRanges. Provided for information for developers, to make it easier to tell (when using a Debugger, or printing debug informatino) which LoopRanges instance is active in looping.loopranges at any point in the code. rank0::Bool Is this process the one with rank 0 in the 'block' which work in parallel on shared memory arrays. <d>::UnitRange{mkint} Loop ranges for each dimension <d> in looping.alldimensions.

moment_kinetics.looping.begin_r_region — Method

Begin region in which (:r,) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_r_vpa_region — Method

Begin region in which (:r, :vpa) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_r_z_region — Method

Begin region in which (:r, :z) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_r_z_vpa_region — Method

Begin region in which (:r, :z, :vpa) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_r_region — Method

Begin region in which (:s, :r) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_r_vpa_region — Method

Begin region in which (:s, :r, :vpa) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_r_z_region — Method

Begin region in which (:s, :r, :z) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_r_z_vpa_region — Method

Begin region in which (:s, :r, :z, :vpa) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_region — Method

Begin region in which (:s,) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_vpa_region — Method

Begin region in which (:s, :vpa) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_z_region — Method

Begin region in which (:s, :z) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_s_z_vpa_region — Method

Begin region in which (:s, :z, :vpa) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_serial_region — Method

Begin region in which only rank-0 in each group of processes operating on a shared-memory block operates on shared-memory arrays.

Returns immediately if loopranges[] is already set for a serial region. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_vpa_region — Method

Begin region in which (:vpa,) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_z_region — Method

Begin region in which (:z,) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.begin_z_vpa_region — Method

Begin region in which (:z, :vpa) dimensions are parallelized by being split between processes.

Returns immediately if loopranges[] is already set to the parallel dimensions being requested. This allows the begin*_region() calls to be placed where they make logical sense, with no cost if a call happens to be repeated (e.g. in different functions).

Calls _block_synchronize() to synchronize the processes operating on a shared-memory block, unless no_synchronize=true is passed as an argument.

moment_kinetics.looping.dims_string — Method

Construct a string composed of the dimension names given in the Tuple dims, separated by underscores

moment_kinetics.looping.get_best_ranges — Method

Find the ranges for loop variables that optimize load balance for a certain block_size

moment_kinetics.looping.get_best_ranges_from_sizes — Method

moment_kinetics.looping.get_load_balance — Method

Calculate the expected load balance

'Load balance' is the ratio of the maximum and minimum numbers of points on any process.

Arguments

nprocslist : Vector{mkint} Number of processes for each dimension sizes : Vector{mk_int} Size of each dimension

moment_kinetics.looping.get_local_range — Method

Get local range of indices when splitting a loop over processes in a sub-block

moment_kinetics.looping.get_splits — Method

Find possible divisions of subblocksize into n factors

moment_kinetics.looping.setup_loop_ranges! — Method

Create ranges for loops with different combinations of variables

Arguments

Keyword arguments dim=n are required for each dim in (:s, :r, :z, :vpa) where n is an integer giving the size of the dimension.

moment_kinetics.looping.@loop_r — Macro

Loop over (:r,) dimensions

moment_kinetics.looping.@loop_r_vpa — Macro

Loop over (:r, :vpa) dimensions

moment_kinetics.looping.@loop_r_z — Macro

Loop over (:r, :z) dimensions

moment_kinetics.looping.@loop_r_z_vpa — Macro

Loop over (:r, :z, :vpa) dimensions

moment_kinetics.looping.@loop_s — Macro

Loop over (:s,) dimensions

moment_kinetics.looping.@loop_s_r — Macro

Loop over (:s, :r) dimensions

moment_kinetics.looping.@loop_s_r_vpa — Macro

Loop over (:s, :r, :vpa) dimensions

moment_kinetics.looping.@loop_s_r_z — Macro

Loop over (:s, :r, :z) dimensions

moment_kinetics.looping.@loop_s_r_z_vpa — Macro

Loop over (:s, :r, :z, :vpa) dimensions

moment_kinetics.looping.@loop_s_vpa — Macro

Loop over (:s, :vpa) dimensions

moment_kinetics.looping.@loop_s_z — Macro

Loop over (:s, :z) dimensions

moment_kinetics.looping.@loop_s_z_vpa — Macro

Loop over (:s, :z, :vpa) dimensions

moment_kinetics.looping.@loop_vpa — Macro

Loop over (:vpa,) dimensions

moment_kinetics.looping.@loop_z — Macro

Loop over (:z,) dimensions

moment_kinetics.looping.@loop_z_vpa — Macro

Loop over (:z, :vpa) dimensions

moment_kinetics.looping.@serial_region — Macro

Run a block of code on only rank-0 of each group of processes operating on a shared-memory block