moment_kinetics_input

moment_kinetics.moment_kinetics_input — Module

Base.get — Method

Utility method for converting a string to an Enum when getting from a Dict, based on the type of the default value

moment_kinetics.moment_kinetics_input.check_coordinate_input — Method

Check input for a coordinate

moment_kinetics.moment_kinetics_input.check_input — Method

check various input options to ensure they are all valid/consistent

moment_kinetics.moment_kinetics_input.check_input_initialization — Method

moment_kinetics.moment_kinetics_input.check_input_time_advance — Method

moment_kinetics.moment_kinetics_input.load_defaults — Method

moment_kinetics.moment_kinetics_input.mk_input — Function

moment_kinetics.moment_kinetics_input.read_input_file — Method

Read input from a TOML file