post_processing

moment_kinetics.post_processing — Module

moment_kinetics.post_processing.allocate_global_zr_fields — Method

functions to allocate arrays that are used at run-time to postprocess data that is stored in the netcdf files

moment_kinetics.post_processing.analyze_and_plot_data — Method

analyze_and_plot_data(prefix...; run_index=nothing)

Make some plots for the simulation at prefix. If more than one argument is passed to prefix, plot all the simulations together.

The strings passed to prefix should be either a directory (which contains run output) or the prefix of output files, like <directory>/<prefix> where the output files are <directory>/<prefix>.moments.h5 and <directory>/<prefix>.dfns.h5.

If a single value is passed for prefix the plots/movies are created in the same directory as the run, and given names based on the name of the run. If multiple values are passed, the plots/movies are given names beginning with compare_ and are created in the comparison_plots/ subdirectory.

By default plots output from all restarts in a directory. To select a single run, pass the run_index argument - the value corresponds to the _<i> suffix given to output files by restart_moment_kinetics(). run_index can be an integer (which is applied to all directories in prefix...), or a tuple of integers (which should have the same length as the number of directories passed to prefix...). Use run_index=-1 to get the most recent run (which does not have a _<i> suffix). Note that run_index is only used when a directory (rather than the prefix of a specific output file) is passed to prefix...

moment_kinetics.post_processing.calculate_and_write_frequencies — Method

moment_kinetics.post_processing.calculate_differences — Method

Find the maximum difference, as a function of time, between two or more outputs for each variable.

moment_kinetics.post_processing.compare_fields_symbolic_test — Method

moment_kinetics.post_processing.draw_v_parallel_zero! — Method

Add a thin, red, dashed line showing vparallel=(vth*wparallel+upar)=0 to a 2d plot with axes (z,vpa).

moment_kinetics.post_processing.fit_cosine — Function

Fit a cosine to a 1d array

Fit function is Acos(2πn(z + δ)/L)

The domain z is taken to be periodic, with the first and last points identified, so L=z[end]-z[begin]

Arguments

z : Array 1d array with positions of the grid points - should have the same length as data data : Array 1d array of the data to be fit amplitudeguess : Float Initial guess for the amplitude (the value from the previous time point might be a good choice) offsetguess : Float Initial guess for the offset (the value from the previous time point might be a good choice) n : Int, default 1 The periodicity used for the fit

Returns

amplitude : Float The amplitude A of the cosine fit offset : Float The offset δ of the cosine fit error : Float The RMS of the difference between data and the fit

moment_kinetics.post_processing.fit_delta_phi_mode — Method

Fit delta_phi to get the frequency and growth rate.

Note, expect the input to be a standing wave (as simulations are initialised with just a density perturbation), so need to extract both frequency and growth rate from the time-variation of the amplitude.

The function assumes that if the amplitude does not cross zero, then the mode is non-oscillatory and so fits just an exponential, not exp*cos. The simulation used as input should be long enough to contain at least ~1 period of oscillation if the mode is oscillatory or the fit will not work.

Arguments

z : Array{mkfloat, 1} 1d array of the grid point positions t : Array{mkfloat, 1} 1d array of the time points deltaphi : Array{mkfloat, 2} 2d array of the values of delta_phi(z, t)

Returns

phifitresult struct whose fields are: growthrate : mkfloat Fitted growth rate of the mode amplitude0 : mkfloat Fitted amplitude at t=0 frequency : mkfloat Fitted frequency of the mode offset0 : mkfloat Fitted offset at t=0 amplitudefiterror : mkfloat RMS error in fit to ln(amplitude) - i.e. ln(A) offsetfiterror : mkfloat RMS error in fit to offset - i.e. δ cosinefiterror : mkfloat Maximum of the RMS errors of the cosine fits at each time point amplitude : Array{mkfloat, 1} Values of amplitude from which growthrate fit was calculated offset : Array{mk_float, 1} Values of offset from which frequency fit was calculated

moment_kinetics.post_processing.get_tuple_of_return_values — Method

get_tuple_of_return_values(func, arg_tuples...)

Suppose func(args...) returns a tuple of return values, then get_tuple_of_return_values(func, arg_tuples...) returns a tuple (with an entry for each return value of func) of tuples (one for each argument in each of arg_tuples...) of return values.

moment_kinetics.post_processing.get_unnormalised_f_coords_2d — Method

Get the unnormalised distribution function and unnormalised ('lab space') coordinates.

Inputs should depend only on z and vpa.

moment_kinetics.post_processing.get_unnormalised_f_dzdt_1d — Method

Get the unnormalised distribution function and unnormalised ('lab space') dzdt coordinate at a point in space.

Inputs should depend only on vpa.

moment_kinetics.post_processing.init_postprocessing_options — Method

moment_kinetics.post_processing.moving_average — Method

Calculate a moving average

result[i] = mean(v[i-n:i+n])

Except near the ends of the array where indices outside the range of v are skipped.

moment_kinetics.post_processing.plot_1D_1V_diagnostics — Method

moment_kinetics.post_processing.plot_charged_pdf — Method

plots various slices of the ion pdf (1d and 2d, stills and animations)

moment_kinetics.post_processing.plot_dfns — Method

moment_kinetics.post_processing.plot_fields — Method

moment_kinetics.post_processing.plot_moments — Method

moment_kinetics.post_processing.plot_neutral_pdf — Method

plots various slices of the neutral pdf (1d and 2d, stills and animations)

moment_kinetics.post_processing.plot_unnormalised_f2d — Method

Make a 2d plot of an unnormalised f on unnormalised coordinates, as returned from getunnormalisedf_coords()

Note this function requires using the PyPlot backend to support 2d coordinates being passed to heatmap().

moment_kinetics.post_processing.read_distributed_zr_data! — Method

Read data

run_names is a tuple. If it has more than one entry, this means that there are multiple restarts (which are sequential in time), so concatenate the data from each entry together.

moment_kinetics.post_processing.trygif — Method

Call gif, but catch the exception if there is an error

moment_kinetics.post_processing.trysavefig — Method

Call savefig, but catch the exception if there is an error