semi_lagrange

moment_kinetics.semi_lagrange — Module

moment_kinetics.semi_lagrange.semi_lagrange_info — Type

structure semilagrangeinfo contains the basic information needed to project backwards along approximate characteristics, which underpins the semi-Lagrange approach to time advancement

moment_kinetics.semi_lagrange.calculate_time_from_boundary! — Method

if periodic boundary condition, will be most efficient to store the time taken by a characteristic to pass from the downwind boundary to each point on the grid, as this will be used for all characteristics that wrap around from -L/2 to L/2

tbound is the time spent on a characteristic in getting from the downdiwnd boundary to the ith grid point

moment_kinetics.semi_lagrange.departure_point! — Method

calculate the departure point for the ith characteristic; overwrites SL.deppts[i], SL.depidx[i], and returns ... note that the dep_idx calculated here is the index of the nearest downwind gridpoint to the departure point (which is in general off-grid)

moment_kinetics.semi_lagrange.departure_point_single_transit! — Method

t_in is the time spent by a particle following this ith characteristic in going from the arrival point (grid point i at future time level) to the grid point jstart; note that the jstart grid point is the grid point immediately downwind of the departure point for the characteristic immediately downwind of this (ith) one.

moment_kinetics.semi_lagrange.find_approximate_characteristic! — Method

moment_kinetics.semi_lagrange.find_departure_points! — Method

follow trajectory from each grid point at time level n+1 backwards in time until Δt has elapsed to find the departure points and the indices of the nearest downwind grid points to the departure points SL.deppts: array containing the locations of the departure points SL.depidx: array containing the indices of the nearest downwind grid point to the departure point tbound: array used to store the cumulative time spent by a particle moving along its characteristic backwards from the downwind boundary. only needed if BC is periodic SL.crossing_time: array containing the amount of time required to cross each cell speed: the advection speed at the grid point locations given by coord.grid coord.grid: grid point locations dt: time step size

moment_kinetics.semi_lagrange.project_characteristics_onto_grid! — Method

determine the nearest grid point to each departure point

moment_kinetics.semi_lagrange.setup_semi_lagrange — Method

create and return a structure containing the arrays needed for the semi-Lagrange time advance

moment_kinetics.semi_lagrange.update_crossing_times! — Method

obtain the time needed to cross the cell assigned to each grid point, given the advection speed within the cell d is an array contain cell widths, with d[1]=d[n] the distance between the 1st and 2nd grid points v is an array containing the speeds at grid points NB: assuming constant advection speed within each cell, taken as speed NB: at neighboring downwind grid point NB: might be better and not much harder to analytically solve NB: for crossing time assuming, e.g., linear variation of advection speed