Setup for moment_kinetics on known clusters

This subdirectory provides scripts to set up Julia and moment_kinetics to run on some known clusters.

Currently supported:

"archer" (ARCHER2)

"marconi" (the EUROfusion supercomputer Marconi)

Quickstart

From the top-level of the moment_kinetics repo, run

$ machines/machine_setup.sh [<path-to-Julia-executable>]

If you omit the <path-to-Julia-executable> argument, you will be prompted to enter it, with a default taken from your $PATH. If there is no julia in your $PATH you will be asked if you want the script to download Julia for you, or you can force the download by passing the -d flag (and do not pass a <path-to-Julia-executable>).

The script will prompt you for several settings, with sensible defaults (where possible). Note that some settings are needed on all machines - if a setting is not needed it will be ignored.

You will be prompted to enter a location for your .julia directory. If you are installing on a cluster which allows access to your home directory from compute nodes, it is fine to leave this as the default. If not (e.g. on ARCHER2), you need to set a path which is accessible from the compute nodes. If you want to create a completely self-contained install (e.g. for reproducibility or for debugging some dependency conflicts), you might want to put .julia within the moment_kinetics directory (e.g. enter . at the prompt).

After editing source code

If you use the precompiled moment_kinetics.so system image, you need to recompile it after editing source code (otherwise simulations will continue to use the old source code).

To do this, run

$ ./precompile-submit.sh

and wait for the resulting job to complete.

This step is required if you use the sumbit-run.sh or submit-restart.sh scripts, as these both use the moment_kinetics.so system image.

Running and restarting simulations

Convenience scripts are provided to submit jobs running and post-processing a simulation run or restart.

$ ./submit-run.sh <path to input file>.toml

will submit a job to run a simulation using that input file and (by default) a linked job that will run the post-processing routines (in the serial queue) when the run finishes.

Similarly

$ ./submit-restart.sh <path to input file>.toml

will submit a job to run and post-process a restart using input file. The simulation will restart from the last time point of the previous run (run_moment_kinetics.jl supports more flexibility, but for now you would need to write your own submission script to pass the options needed for that).

Default parameters for the runs (number of nodes, time limit, etc.) were set up by machines/machine_setup.sh are stored in LocalPreferences.toml (which can be edited to change them). The parameters can be altered for a particular job (and you can disable the post-processing job) using command line flags described by the help text

$ ./submit-run.sh -h

or

$ ./submit-restart.sh -h

Advanced usage

The convenience script machine_setup.sh is provide because the actual setup happens in multiple stages, with Julia being restarted in between (as this is required on some machines):

Start Julia (using the Julia executable you want to use in the end for simulations) from the top level of the moment_kinetics repo, but not requiring the --project flag; run include("machines/shared/machine_setup.jl"); call the moment_kinetics.machine_setup_moment_kinetics function with appropriate settings as arguments (see the docstring).

This function requires no packages outside the base Julia system image (because on some machines it is desirable not to have to install any packages before setting JULIA_DEPOT_PATH, which can be done by this function).

It saves some settings into LocalPreferences.toml; makes a symlink at bin/julia to the Julia executable being used; saves environment setup into the julia.env file (which can include setting the JULIA_DEPOT_PATH environment variable to ensure the .julia directory is accessible on compute nodes); and if necessary symlinks a machine-specific file from machines/shared/machine_setup_stage_two.jl.

Usually Julia needs to be restarted after running this function, so it will call exit() to stop Julia. This can be avoided by passing the no_force_exit=true argument if that is useful.

Run source julia.env in each terminal session where you want to use moment_kinetics, or add it to your .bashrc (if this does not conflict with any other projects).

Note that julia.env runs module purge to remove any already loaded modules (to get a clean environment). It is therefore very likely to interfere with other projects.

If it is necessary to run a second stage of setup, for example after Julia is restarted with a special JULIA_DEPOT_PATH, start Julia again, this time with the --project flag. A symlink has been created, so you can use
$ bin/julia --project
and run include("machines/shared/machine_setup_stage_two.jl").

This script sets up MPI and HDF5.

It exits Julia when it is finished, as Julia must be restarted when MPIPreferences settings are changed.

If you want to build a precompiled system image (to speed up startup of simulation runs), run
$ precompile-submit.sh
which will submit a serial or debug job that runs precompile.jl to create the moment_kinetics.so image.

This is required if using the submit-run.sh script, unless you edit the jobscript templates in the appropriate subdirectory machines/<your-machine>.

API documentation

moment_kinetics.machine_setup — Module

Functions to help setting up on known machines

moment_kinetics.machine_setup.machine_setup_moment_kinetics — Method

machine_setup_moment_kinetics(machine::String,
                              account::String,
                              julia_directory::String,
                              default_run_time::String,
                              default_nodes::String,
                              default_postproc_time::String,
                              default_postproc_memory::String,
                              default_partition::String;
                              default_qos::String;
                              no_force_exit::Bool=false,
                              interactive::Bool=true)

Do setup for a known machine:

On clusters that use a module system, provide julia.env at the top level of the moment_kinetics repo.

Call

source julia.env

to get the correct modules for running moment_kinetics, either on the command line (to get them for the current session) or in your .bashrc (to get them by default). Note that this calls module purge so will remove any currently loaded modules when it is run.

Run setup commands for MPI and HDF5 which ensure the correct, system-provided libraries are used.

Makes a symlink to the Julia exeutable used to run this command at bin/julia under the moment_kinetics repo, so that setup and job submission scripts can use a known relative path.

Note

If you change the Julia executable, e.g. to update to a new verison, you will need to either replace the symlink <moment_kinetics>/bin/julia by hand, or re-run this function using the new executable.

julia_directory gives the location of the directory (usually called .julia) where Julia installs files, saves settings, etc. julia_directory must be passed if this directory should be in a non-default location (i.e. not $HOME/.julia/). The value is used to set JULIA_DEPOT_PATH in the julia.env file, so that this setting is propagated to the environment on the compute nodes.

Usually it is necessary for Julia to be restarted after running this function to ensure the correct MPI is linked, etc. so the function will force Julia to exit. If for some reason this is not desired (e.g. when debugging), pass no_force_exit=true.

The interactive argument exists so that when this function is called from another script, terminal output with instructions for the next step can be disabled.

The remaining arguments can be used to change the default settings for jobs submitted using the provided submit-run.sh script. These settings are read by the scripts from LocalPreferences.toml and the values can safely be edited in that file without re-running this function (if you want to). The arguments are:

default_run_time is the maximum run time for the simulation, in the format expected by sbatch --time, e.g. "24:00:00" for 24 hours, 0 minutes, 0 seconds.

default_nodes is the default number of nodes to use for a simulation run. Note that post-processing always runs in serial (using a serial or debug queue if available).

default_postproc_time is the maximum run time for the post-processing job, in the format expected by sbatch --time, e.g. "1:00:00" for 1 hours, 0 minutes, 0 seconds.

default_postproc_memory is the memory requested for the post-processing job, in the format expected by sbatch --mem, e.g. "64G" for 64GB.

default_partition is the default 'partition' passed to sbatch --partition. See your cluster's documentation for possible values. The default will be the standard queue, which charges towards the budget of your allocation. You might sometimes want, for example, to change this to a debug queue if one is available.

default_qos is the default 'quality of service' passed to sbatch --qos. See your cluster's documentation for possible values. The default will be the standard queue, which charges towards the budget of your allocation. You might want, for example, to change this to a free, low-priority queue if one is available.

Currently supported machines:

"archer" - the UK supercomputer ARCHER2

"marconi" - the EUROfusion supercomputer Marconi

Note

The settings created by this function are saved in LocalPreferences.toml (using the Preferences.jl package). It might sometimes be useful to edit these by hand (e.g. the account setting if this needs to be changed.): it is fine to do this.