source_terms

moment_kinetics.source_terms — Module

moment_kinetics.source_terms.source_terms! — Method

calculate the source terms due to redefinition of the pdf to split off density, flow and/or pressure, and use them to update the pdf

moment_kinetics.source_terms.source_terms_evolve_density! — Method

moment_kinetics.source_terms.source_terms_evolve_density_neutral! — Method

moment_kinetics.source_terms.source_terms_evolve_ppar_collisions! — Method

update the evolved pdf to account for the charge exchange and ionization source terms in the kinetic equation arising due to the re-normalization of the pdf as g = f * vth / n

moment_kinetics.source_terms.source_terms_evolve_ppar_collisions_neutral! — Method

update the evolved pdf to account for the charge exchange and ionization source terms in the kinetic equation arising due to the re-normalization of the pdf as g = f * vth / n

moment_kinetics.source_terms.source_terms_evolve_ppar_no_collisions! — Method

update the evolved pdf to account for the collisionless source terms in the kinetic equation arising due to the re-normalization of the pdf as g = f * vth / n

moment_kinetics.source_terms.source_terms_evolve_ppar_no_collisions_neutral! — Method

update the evolved pdf to account for the collisionless source terms in the kinetic equation arising due to the re-normalization of the pdf as g = f * vth / n

moment_kinetics.source_terms.source_terms_manufactured! — Method

advance the dfn with an arbitrary source function

moment_kinetics.source_terms.source_terms_neutral! — Method

calculate the source terms due to redefinition of the pdf to split off density, flow and/or pressure, and use them to update the pdf