time_advance

moment_kinetics.time_advance — Module

moment_kinetics.time_advance.advance_info — Type

moment_kinetics.time_advance.euler_time_advance! — Method

eulertimeadvance! advances the vector equation dfvec/dt = G[f] that includes the kinetic equation + any evolved moment equations using the forward Euler method: fvecout = fvecin + dt*fvecin, with fvecin an input and fvec_out the output

moment_kinetics.time_advance.normalize_pdf! — Method

if evolving the density via continuity equation, redefine the normalised f → f/n if evolving the parallel pressure via energy equation, redefine f -> f * vth / n 'scratch' should be a (nz,nspecies) array

moment_kinetics.time_advance.rk_update! — Method

moment_kinetics.time_advance.setup_runge_kutta_coefficients — Method

given the number of Runge Kutta stages that are requested, returns the needed Runge Kutta coefficients; e.g., if f is the function to be updated, then f^{n+1}[stage+1] = rkcoef[1,stage]*f^{n} + rkcoef[2,stage]f^{n+1}[stage] + rk_coef[3,stage](f^{n}+dt*G[f^{n+1}[stage]]

moment_kinetics.time_advance.setup_scratch_arrays — Method

create an array of structs containing scratch arrays for the normalised pdf and low-order moments that may be evolved separately via fluid equations

moment_kinetics.time_advance.setup_time_advance! — Method

create arrays and do other work needed to setup the main time advance loop. this includes creating and populating structs for Chebyshev transforms, velocity space moments, EM fields, semi-Lagrange treatment, and advection terms

moment_kinetics.time_advance.ssp_rk! — Method

moment_kinetics.time_advance.time_advance! — Method

solve ∂f/∂t + v(z,t)⋅∂f/∂z + dvpa/dt ⋅ ∂f/∂vpa= 0 define approximate characteristic velocity v₀(z)=vⁿ(z) and take time derivative along this characteristic df/dt + δv⋅∂f/∂z = 0, with δv(z,t)=v(z,t)-v₀(z) for prudent choice of v₀, expect δv≪v so that explicit time integrator can be used without severe CFL condition

moment_kinetics.time_advance.time_advance_no_splitting! — Method

moment_kinetics.time_advance.time_advance_split_operators! — Method

moment_kinetics.time_advance.update_pdf_unnorm! — Method

if separately evolving the density via the continuity equation, the evolved pdf has been normalised by the particle density undo this normalisation to get the true particle distribution function

scratch should be a (nz,nspecies) array

moment_kinetics.time_advance.update_solution_vector! — Method

update the vector containing the pdf and any evolved moments of the pdf for use in the Runge-Kutta time advance