finite_differences

moment_kinetics.finite_differences — Module

moment_kinetics.finite_differences.finite_difference_info — Type

Finite difference discretization

moment_kinetics.calculus.elementwise_derivative! — Method

elementwise_derivative!(coord, f, adv_fac, not_spectral::finite_difference_info)

Calculate the derivative of f using finite differences, with particular scheme specified by coord.fdoption; result stored in coord.scratch2d.

moment_kinetics.calculus.elementwise_derivative! — Method

elementwise_derivative!(coord, f, not_spectral::finite_difference_info)

Calculate the derivative of f using 4th order centered finite differences; result stored in coord.scratch_2d.

moment_kinetics.calculus.elementwise_second_derivative! — Method

elementwise_second_derivative!(coord, f, not_spectral::finite_difference_info)

Calculate the second derivative of f using 2nd order centered finite differences; result stored in coord.scratch_2d.

moment_kinetics.finite_differences.centered_fourth_order! — Method

take the derivative of input function f and return as df using fourth-order, centered differences. input/output array df is 2D array of size ngrid x nelement

moment_kinetics.finite_differences.centered_second_order! — Method

take the derivative of input function f and return as df using second-order, centered differences. input/output array df is 2D array of size ngrid x nelement

moment_kinetics.finite_differences.centered_second_order! — Method

take the derivative of input function f and return as df using second-order, centered differences. input/output df is 1D array of size n (full grid)

moment_kinetics.finite_differences.derivative_finite_difference! — Method

moment_kinetics.finite_differences.derivative_finite_difference! — Method

moment_kinetics.finite_differences.fd_check_option — Method

moment_kinetics.finite_differences.second_derivative_finite_difference! — Method

Take the second derivative of input function f and return as df using second-order, centered differences. output array df is 2D array of size ngrid x nelement

moment_kinetics.finite_differences.upwind_first_order! — Method

moment_kinetics.finite_differences.upwind_second_order! — Method

moment_kinetics.finite_differences.upwind_third_order! — Method