gauss_legendre

moment_kinetics.gauss_legendre — Module

module for Gauss-Legendre-Lobatto and Gauss-Legendre-Radau spectral element grids

moment_kinetics.gauss_legendre.gausslegendre_base_info — Type

A struct for passing around elemental matrices on Gauss-Legendre points in 1D

moment_kinetics.gauss_legendre.gausslegendre_info — Type

A struct for Gauss-Legendre arrays needed for global operations in 1D, contains the struct of elemental matrices for Lobatto and Radau points, as well as some assembled 1D global matrices.

moment_kinetics.calculus.elementwise_derivative! — Method

A function that takes the first derivative in each element of coord.grid, leaving the result (element-wise) in coord.scratch_2d.

moment_kinetics.calculus.mass_matrix_solve! — Method

Function to carry out a 1D (global) mass matrix solve.

moment_kinetics.gauss_legendre.GaussLegendre_derivative_vector! — Method

Gauss-Legendre derivative at arbitrary x values, for boundary condition on Radau points. D0 – the vector xj – the x location where the derivative is evaluated ngrid – number of points in x x – the grid from -1, 1 Note that D0 is not scaled to the physical grid with a scaling factor.

moment_kinetics.gauss_legendre.Legendre_h_n — Method

Result of the inner product of Legendre polynomials of order k.

moment_kinetics.gauss_legendre.gausslobattolegendre_differentiation_matrix! — Method

Formula for Gauss-Legendre-Lobatto differentiation matrix taken from p196 of Chpt The Spectral Elemtent Method' ofComputational Seismology'. Heiner Igel First Edition. Published in 2017 by Oxford University Press. Or https://doc.nektar.info/tutorials/latest/fundamentals/differentiation/fundamentals-differentiationch2.html

D -- differentiation matrix 
x -- Gauss-Legendre-Lobatto points in [-1,1]
ngrid -- number of points per element (incl. boundary points)

Note that D has does not include a scaling factor

moment_kinetics.gauss_legendre.gaussradaulegendre_differentiation_matrix! — Method

Formula for Gauss-Legendre-Radau differentiation matrix taken from https://doc.nektar.info/tutorials/latest/fundamentals/differentiation/fundamentals-differentiationch2.html

D -- differentiation matrix 
x -- Gauss-Legendre-Radau points in [-1,1)
ngrid -- number of points per element (incl. boundary points)

Note that D has does not include a scaling factor

moment_kinetics.gauss_legendre.get_KK_local! — Method

If called for coord.name = vperp elemental matrix KK on the $i^{th}$ element is

\[ K_{jk} = -\int^{v_\perp^U}_{v_\perp^L} \left(v_\perp \frac{\partial\varphi_j(v_\perp)}{\partial v_\perp} + \varphi_j(v_\perp) \right) \frac{\partial\varphi_k(v_\perp)}{\partial v_\perp} d v_\perp = -\int^1_{-1} ((c_i + x s_i)l_j^\prime(x) + l_j(x))l_k^\prime(x) d x /s_i\]

with $c_i$ and $s_i$ the appropriate shift and scale factors, respectively. Otherwise, if called for any other coordinate elemental matrix KK is the same as LL (see get_LL_local!()). IfexplicitBCterms = true`, boundary terms arising from integration by parts are included at the extreme boundary points.

moment_kinetics.gauss_legendre.get_LL_local! — Method

If called for coord.name = vperp elemental matrix LL on the $i^{th}$ element is

\[ L_{jk} = -\int^{v_\perp^U}_{v_\perp^L} \frac{\partial\varphi_j(v_\perp)}{\partial v_\perp}\frac{\partial\varphi_k(v_\perp)}{\partial v_\perp} v_\perp d v_\perp = -\int^1_{-1} (c_i + x s_i)l_j^\prime(x)l_k^\prime(x) d x /s_i\]

with $c_i$ and $s_i$ the appropriate shift and scale factors, respectively. Otherwise, if called for any other coordinate elemental matrix LL is

\[ L_{jk} = -\int^{v_\|^U}_{v_\|^L} \frac{\partial\varphi_j(v_\|)}{\partial v_\|}\frac{\partial\varphi_k(v_\|)}{\partial v_\|} d v_\| = -\int^1_{-1} l_j^\prime(x)l_k^\prime(x) d x /s_i.\]

If explicit_BC_terms = true, boundary terms arising from integration by parts are included at the extreme boundary points.

moment_kinetics.gauss_legendre.get_QQ_local! — Method

Construction function to provide the appropriate elemental matrix Q to the global matrix assembly functions.

moment_kinetics.gauss_legendre.integration_matrix! — Method

Function to calculate the elemental anti-differentiation (integration) matrix. This function forms the primitive

\[F(x) = \int^x_{x_{\rm min}} f(x^\prime) d x^\prime\]

of the function

\[f(x) = \sum_j f_j l_j(x),\]

where $l_j(x)$ is the $j^{\rm th}$ Lagrange polynomial on the element and $f_j = f(x_j)$, with $x_j$ $j^{\rm th}$ collocation point on the element. We find $F(x)$ at the collocation points on the element, giving a series of integrals to evaluate:

\[F(x_i) = \int^{x_i}_{-1} f(x^\prime) d x^\prime = \sum_j f_j \int^{x_i}_{-1} l_j(x^\prime) d x^\prime,\]

where we have used that $x_{\rm min} = -1$ on the elemental grid. Changing to a normalised coordinate $y$ suitable for Gaussian quadrature

\[x^\prime = \frac{x_i + 1}{2} y + \frac{x_i - 1}{2}\]

we can write the operation in matrix form:

\[F(x_i) = \sum_{j}A_{ij}f_j,\]

with the matrix $A_{ij}$ defined by

\[A_{ij} = \left(\frac{x_i + 1}{2}\right) \int^1_{-1} l_j \left( \frac{(x_i + 1)y + x_i - 1}{2} \right) dy,\]

or in discretised form

\[A_{ij} = \left(\frac{x_i + 1}{2}\right) \sum_k l_j \left( \frac{(x_i + 1)y_k + x_i - 1}{2} \right) w_k,\]

with $y_k$ and $w_k$ Gauss-quadrature points and weights, respectively.

moment_kinetics.gauss_legendre.scaled_gauss_legendre_lobatto_grid — Method

Function for setting up the full Gauss-Legendre-Lobatto grid and collocation point weights.

moment_kinetics.gauss_legendre.scaled_gauss_legendre_radau_grid — Method

Function for setting up the full Gauss-Legendre-Radau grid and collocation point weights.

moment_kinetics.gauss_legendre.setup_gausslegendre_pseudospectral — Method

Function to create gausslegendre_info struct.

moment_kinetics.gauss_legendre.setup_gausslegendre_pseudospectral_lobatto — Method

Function that creates the gausslegendre_base_info struct for Lobatto points.

moment_kinetics.gauss_legendre.setup_gausslegendre_pseudospectral_radau — Method

Function that creates the gausslegendre_base_info struct for Radau points.

moment_kinetics.gauss_legendre.setup_global_strong_form_matrix! — Method

A function that assigns the local matrices to a global array QQ_global for later evaluating strong form of required 1D equation.

The 'option' variable is a flag for choosing the type of matrix to be constructed. Currently the function is set up to assemble the elemental matrices without imposing boundary conditions on the first and final rows of the matrix. This means that the operators constructed from this function can only be used for differentiation, and not solving 1D ODEs. The shared points in the element assembly are averaged (instead of simply added) to be consistent with the derivative_elements_to_full_grid!() function in calculus.jl.

moment_kinetics.gauss_legendre.setup_global_weak_form_matrix! — Method

A function that assigns the local weak-form matrices to a global array QQ_global for later solving weak form of required 1D equation.

The 'option' variable is a flag for choosing the type of matrix to be constructed. Currently the function is set up to assemble the elemental matrices without imposing boundary conditions on the first and final rows of the matrix by default. This means that the operators constructed from this function can only be used for differentiation, and not solving 1D ODEs. This assembly function assumes that the coordinate is not distributed. To extend this function to support distributed-memory MPI, addition of off-memory matrix elements to the exterior points would be required.

The typical use of this function is to assemble matrixes M and K in

M * d2f = K * f

where M is the mass matrix and K is the stiffness matrix, and we wish to solve for d2f given f. To solve 1D ODEs

K * f = b = M * d2f

for f given boundary data on f with periodic or dirichlet boundary conditions, set

periodic_bc = true, b[end] = 0

or

dirichlet_bc = true, b[1] = f[1] (except for cylindrical coordinates), b[end] = f[end]

in the function call, and create new matrices for this purpose in the gausslegendre_info struct. Currently the Laplacian matrix is supported with boundary conditions.

moment_kinetics.interpolation.single_element_interpolate! — Method

Function to perform interpolation on a single element.