interpolation

moment_kinetics.interpolation — Module

Interpolation routines intended for post-processing.

Note these are not guaranteed to be highly optimized!

moment_kinetics.interpolation.interpolate_to_grid_1d! — Function

Interpolation from a regular grid to a 1d grid with arbitrary spacing

Arguments

result : Array{mkfloat, 1} Array to be overwritten with the result of the interpolation newgrid : Array{mkfloat, 1} Grid of points to interpolate coord to f : Array{mkfloat} Field to be interpolated coord : coordinate coordinate struct giving the coordinate along which f varies spectral : discretization_info struct containing information for discretization, whose type determines which method is used.

moment_kinetics.interpolation.interpolate_to_grid_1d — Method

Interpolation from a regular grid to a 1d grid with arbitrary spacing

This version allocates a new array for the result, which is returned.

Arguments

newgrid : Array{mkfloat, 1} Grid of points to interpolate coord to f : Array{mkfloat} Field to be interpolated coord : coordinate coordinate struct giving the coordinate along which f varies spectral : Bool or chebyshevinfo struct containing information for discretization, whose type determines which method is used.

Returns

result : Array Array with the values of f interpolated to the points in new_grid.

moment_kinetics.interpolation.interpolate_to_grid_vpa! — Method

moment_kinetics.interpolation.interpolate_to_grid_vpa! — Method

moment_kinetics.interpolation.interpolate_to_grid_vpa — Method

moment_kinetics.interpolation.interpolate_to_grid_vpa — Method

moment_kinetics.interpolation.interpolate_to_grid_z! — Method

moment_kinetics.interpolation.interpolate_to_grid_z! — Method

moment_kinetics.interpolation.interpolate_to_grid_z! — Method

moment_kinetics.interpolation.interpolate_to_grid_z — Method

moment_kinetics.interpolation.interpolate_to_grid_z — Method

moment_kinetics.interpolation.interpolate_to_grid_z — Method