moment_kinetics

moment_kinetics.moment_kinetics — Module

moment_kinetics.check_so_newer_than_code — Function

check_so_newer_than_code(so_filename=nothing)

Utility function that checks if so_filename is newer than the source code in moment_kinetics/src. If it is, prints an error message and returns false; otherwise returns true.

If so_filename is nothing, use the name of the system image of the current julia session for so_filename.

If so_filename is "makie_postproc.so", also checks against the source code in makie_post_processing/makie_post_processing/src/.

If so_filename is "plots_postproc.so", also checks against the source code in plots_post_processing/plots_post_processing/src/.

moment_kinetics.cleanup_moment_kinetics! — Method

Clean up after a run

moment_kinetics.get_backup_filename — Method

Append a number to the filename, to get a new, non-existing filename to backup the file to.

moment_kinetics.run_moment_kinetics — Function

main function that contains all of the content of the program

moment_kinetics.run_moment_kinetics — Method

overload with no TimerOutput arguments

moment_kinetics.run_moment_kinetics — Method

overload which takes a filename and loads input

moment_kinetics.run_moment_kinetics — Method

overload which gets the input file name from command line arguments

moment_kinetics.setup_moment_kinetics — Method

Perform all the initialization steps for a run.

If backup_filename is nothing, set up for a regular run; if a filename is passed, reload data from time index given by restart_time_index for a restart.

debug_loop_type and debug_loop_parallel_dims are used to force specific set ups for parallel loop ranges, and are only used by the tests in debug_test/.