utils

moment_kinetics.utils — Module

Utility functions

moment_kinetics.utils.get_CFL — Function

get_CFL!(CFL, speed, coord)

Calculate the CFL factor 'speed/(grid spacing)' (with no prefactor) corresponding to advection speed speed for advection. Note that moment_kinetics is set up so that dimension in which advection happens is the first dimension of speed - coord is the coordinate corresponding to this dimension.

The result is written in CFL. This function is only intended to be used in post-processing.

moment_kinetics.utils.get_minimum_CFL_neutral_vz — Method

get_minimum_CFL_neutral_vz(speed, vz)

Calculate the minimum (over a shared-memory block) of the CFL factor 'speed/(grid spacing)' (with no prefactor) corresponding to advection speed speed for advection of neutrals in the vz direction.

Reduces the result over the shared-memory block (handling distributed parallelism is left to the calling site). The result is only to be used on rank-0 of the shared-memory block.

moment_kinetics.utils.get_minimum_CFL_neutral_z — Method

get_minimum_CFL_neutral_z(speed, z)

Calculate the minimum (over a shared-memory block) of the CFL factor 'speed/(grid spacing)' (with no prefactor) corresponding to advection speed speed for advection of neutrals in the z direction.

Reduces the result over the shared-memory block (handling distributed parallelism is left to the calling site). The result is only to be used on rank-0 of the shared-memory block.

moment_kinetics.utils.get_minimum_CFL_vpa — Method

get_minimum_CFL_vpa(speed, vpa)

Calculate the minimum (over a shared-memory block) of the CFL factor 'speed/(grid spacing)' (with no prefactor) corresponding to advection speed speed for advection in the vpa direction.

Reduces the result over the shared-memory block (handling distributed parallelism is left to the calling site). The result is only to be used on rank-0 of the shared-memory block.

moment_kinetics.utils.get_minimum_CFL_z — Method

get_minimum_CFL_z(speed, z)

Calculate the minimum (over a shared-memory block) of the CFL factor 'speed/(grid spacing)' (with no prefactor) corresponding to advection speed speed for advection in the z direction.

Reduces the result over the shared-memory block (handling distributed parallelism is left to the calling site). The result is only to be used on rank-0 of the shared-memory block.

moment_kinetics.utils.get_unnormalized_parameters — Function

get_unnormalized_parameters(input::Dict)
get_unnormalized_parameters(input_filename::String)

Get many parameters for the simulation setup given by input or in the file input_filename, in SI units and eV, returned as an OrderedDict.

moment_kinetics.utils.print_unnormalized_parameters — Method

print_unnormalized_parameters(input)

Print many parameters for the simulation setup given by input (a Dict of parameters or a String giving a filename), in SI units and eV.

moment_kinetics.utils.to_hours — Method

to_hours(x::T) where {T<:TimePeriod}

Convert a time period x to seconds

moment_kinetics.utils.to_minutes — Method

to_minutes(x::T) where {T<:TimePeriod}

Convert a time period x to seconds

moment_kinetics.utils.to_seconds — Method

to_seconds(x::T) where {T<:TimePeriod}

Convert a time period x to seconds