velocity_moments

moment_kinetics.velocity_moments — Module

moment_kinetics.velocity_moments.calculate_electron_moment_derivatives! — Method

Pre-calculate spatial derivatives of the electron moments that will be needed for the time advance

moment_kinetics.velocity_moments.calculate_electron_moment_derivatives_no_r! — Method

Calculate spatial derivatives of the electron moments.

This version, for use in implicit solvers for electrons, works with a single point in r, given by ir.

moment_kinetics.velocity_moments.calculate_ion_moment_derivatives! — Method

Pre-calculate spatial derivatives of the moments that will be needed for the time advance

moment_kinetics.velocity_moments.calculate_ion_qpar_from_coll_krook! — Method

calculate parallel heat flux if ion composition flag is coll_krook fluid ions

moment_kinetics.velocity_moments.calculate_ion_qpar_from_pdf! — Method

calculate parallel heat flux if ion composition flag is kinetic ions

moment_kinetics.velocity_moments.calculate_neutral_moment_derivatives! — Method

Pre-calculate spatial derivatives of the neutral moments that will be needed for the time advance

moment_kinetics.velocity_moments.create_moments_electron — Method

create a moment struct containing information about the electron moments

moment_kinetics.velocity_moments.create_moments_ion — Method

moment_kinetics.velocity_moments.integrate_over_negative_vpa — Method

computes the integral over vpa <= 0 of the integrand, using the input vpa_wgts this could be made more efficient for the case that dz/dt = vpa is time-independent, but it has been left general for the cases where, e.g., dz/dt = wpa*vth + upar varies in time

moment_kinetics.velocity_moments.integrate_over_positive_vpa — Method

computes the integral over vpa >= 0 of the integrand, using the input vpa_wgts this could be made more efficient for the case that dz/dt = vpa is time-independent, but it has been left general for the cases where, e.g., dz/dt = wpa*vth + upar varies in time

moment_kinetics.velocity_moments.reset_moments_status! — Method

moment_kinetics.velocity_moments.update_chodura! — Method

runtime diagnostic routine for computing the Chodura ratio in a single species plasma with Z = 1

moment_kinetics.velocity_moments.update_chodura_integral_species! — Method

compute the integral needed for the generalised Chodura condition

IChodura = (Z^2 vBohm^2 / cref^2) * int ( f bz^2 / vz^2 + dfdr*rhostar/vz ) vBohm = sqrt(Z Te/mi) with Z = 1 and mref = mi cref = sqrt(2Ti/mi) and normalise to the local ion density, appropriate to assessing the Chodura condition

IChodura <= (Te/e)d ne / dphi |(sheath entrance) = ni

to a single species plasma with Z = 1

moment_kinetics.velocity_moments.update_density! — Method

NB: if this function is called and if dens_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_density_species! — Method

calculate the updated density (dens) for a given species; should only be called when evolvedensity = false, in which case the vpa coordinate is vpa/cs

moment_kinetics.velocity_moments.update_derived_electron_moment_time_derivatives! — Method

'Primary' time derivatives are calculated when doing the time advance for moment quantities. Moment kinetic equations require in addition some 'derived' time derivatives, which we can calculate by applying the chain rule.

For electrons the kinetic equation that is solved for the shape function is simplified by using $\sqrt{m_e/m_i}$ as a small parameter after substituting in the moment equations. Therefore it is not possible (or at least not convenient) to use dp_dt as calculated from the moment equation. The electron moment time derivatives may still be useful as diagnostics, so we calculate them anyway, including the derived versions (calculated in this function), similar to the ion moment time derivatives.

Note that due to the implicit/explicit splitting of terms in the timestep, which means that the density (which is equal to the ion density) does not update during the implicit electron timestep, the time derivative of density dn/dt does not contribute to the chain rule calculations in this function, even though analytically it would contribute (although the contribution would be small in sqrt(me/mi)). Another way of saying this is that due to the operator splitting, the time derivatives during the implicit step are done with the explicit variables (density in particular) held fixed, and so dvth_dt|_n = 0.5 * vth * dp_dt / p.

moment_kinetics.velocity_moments.update_derived_ion_moment_time_derivatives! — Method

'Primary' time derivatives are calculated when doing the time advance for moment quantities. Moment kinetic equations require in addition some 'derived' time derivatives, which we can calculate by applying the chain rule.

moment_kinetics.velocity_moments.update_derived_moments! — Method

update velocity moments that are calculable from the evolved ion pdf

moment_kinetics.velocity_moments.update_derived_moments_neutral! — Method

update velocity moments that are calculable from the evolved neutral pdf

moment_kinetics.velocity_moments.update_derived_neutral_moment_time_derivatives! — Method

'Primary' time derivatives are calculated when doing the time advance for moment quantities. Moment kinetic equations require in addition some 'derived' time derivatives, which we can calculate by applying the chain rule.

moment_kinetics.velocity_moments.update_ion_qpar! — Method

NB: the incoming pdf is the normalized pdf

moment_kinetics.velocity_moments.update_ion_qpar_species! — Method

calculate the updated parallel heat flux (qpar) for a given species

moment_kinetics.velocity_moments.update_moments! — Method

calculate the updated density (dens) and parallel pressure (ppar) for all species this function is only used once after initialisation the function used to update moments at run time is updatederivedmoments! in time_advance.jl

moment_kinetics.velocity_moments.update_moments_neutral! — Method

update velocity moments of the evolved neutral pdf

moment_kinetics.velocity_moments.update_neutral_density! — Method

calculate the neutral density from the neutral pdf

moment_kinetics.velocity_moments.update_neutral_density_species! — Method

calculate the updated density (dens) for a given species

moment_kinetics.velocity_moments.update_neutral_p_species! — Method

calculate the updated pressure (p) for a given species

moment_kinetics.velocity_moments.update_neutral_pr_species! — Method

calculate the updated pressure in rr direction (pr) for a given species

moment_kinetics.velocity_moments.update_neutral_pz_species! — Method

calculate the updated pressure in zz direction (pz) for a given species

moment_kinetics.velocity_moments.update_neutral_pzeta_species! — Method

calculate the updated pressure in zeta-zeta direction (pzeta) for a given species

moment_kinetics.velocity_moments.update_neutral_qz_species! — Method

calculate the updated heat flux zzz direction (qz) for a given species

moment_kinetics.velocity_moments.update_neutral_ur_species! — Method

calculate the updated ur (mean velocity in r) for a given species

moment_kinetics.velocity_moments.update_neutral_uz_species! — Method

calculate the updated uz (mean velocity in z) for a given species

moment_kinetics.velocity_moments.update_neutral_uzeta_species! — Method

calculate the updated uzeta (mean velocity in zeta) for a given species

moment_kinetics.velocity_moments.update_p! — Method

NB: if this function is called and if p_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_p_species! — Method

calculate the updated energy density (or pressure, p) for a given species; which of these is calculated depends on the definition of the vpa coordinate

moment_kinetics.velocity_moments.update_ppar! — Method

Calculate the parallel pressure p∥=∫d^3v (v∥ - u_∥)^2 f

moment_kinetics.velocity_moments.update_ppar_species! — Method

calculate the updated energy density (or parallel pressure, ppar) for a given species; which of these is calculated depends on the definition of the vpa coordinate

moment_kinetics.velocity_moments.update_pperp_species! — Method

calculate the updated perpendicular pressure (pperp) for a given species

moment_kinetics.velocity_moments.update_upar! — Method

NB: if this function is called and if upar_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_upar_species! — Method

calculate the updated parallel flow (upar) for a given species