vpa_advection

moment_kinetics.vpa_advection — Module

moment_kinetics.vpa_advection.implicit_vpa_advection! — Method

moment_kinetics.vpa_advection.update_speed_vpa! — Method

calculate the advection speed in the vpa-direction at each grid point

moment_kinetics.vpa_advection.update_speed_vpa_inner! — Method

update the advection speed in the parallel velocity coordinate for the case where density, flow and pressure are evolved independently from the pdf; in this case, the parallel velocity coordinate is the normalized peculiar velocity wpa = (vpa - upar)/vth

moment_kinetics.vpa_advection.update_speed_vpa_inner! — Method

update the advection speed in the parallel velocity coordinate for the case where density and flow are evolved independently from the pdf; in this case, the parallel velocity coordinate is the peculiar velocity wpa = vpa-upar

moment_kinetics.vpa_advection.update_speed_vpa_inner! — Method

update the advection speed in the parallel velocity coordinate for the case where no moments are evolved independently from the pdf. vpa is unchanged.

moment_kinetics.vpa_advection.update_speed_vpa_inner! — Method

update the advection speed in the parallel velocity coordinate for the case where density is evolved independently from the pdf; in this case, the parallel velocity coordinate is unchanged.

moment_kinetics.vpa_advection.vpa_advection! — Method