post_processing

moment_kinetics.post_processing — Module

moment_kinetics.post_processing.analyze_and_plot_data — Method

moment_kinetics.post_processing.calculate_and_write_frequencies — Method

moment_kinetics.post_processing.fit_cosine — Function

Fit a cosine to a 1d array

Fit function is Acos(2πn(z + δ)/L)

The domain z is taken to be periodic, with the first and last points identified, so L=z[end]-z[begin]

Arguments

z : Array 1d array with positions of the grid points - should have the same length as data data : Array 1d array of the data to be fit amplitudeguess : Float Initial guess for the amplitude (the value from the previous time point might be a good choice) offsetguess : Float Initial guess for the offset (the value from the previous time point might be a good choice) n : Int, default 1 The periodicity used for the fit

Returns

amplitude : Float The amplitude A of the cosine fit offset : Float The offset δ of the cosine fit error : Float The RMS of the difference between data and the fit

moment_kinetics.post_processing.fit_delta_phi_mode — Method

Fit delta_phi to get the frequency and growth rate.

Note, expect the input to be a standing wave (as simulations are initialised with just a density perturbation), so need to extract both frequency and growth rate from the time-variation of the amplitude.

The function assumes that if the amplitude does not cross zero, then the mode is non-oscillatory and so fits just an exponential, not exp*cos. The simulation used as input should be long enough to contain at least ~1 period of oscillation if the mode is oscillatory or the fit will not work.

Arguments

z : Array{mkfloat, 1} 1d array of the grid point positions t : Array{mkfloat, 1} 1d array of the time points deltaphi : Array{mkfloat, 2} 2d array of the values of delta_phi(z, t)

Returns

phifitresult struct whose fields are: growthrate : mkflaot Fitted growth rate of the mode amplitude0 : mkfloat Fitted amplitude at t=0 frequency : mkfloat Fitted frequency of the mode offset0 : mkfloat Fitted offset at t=0 amplitudefiterror : mkfloat RMS error in fit to ln(amplitude) - i.e. ln(A) offsetfiterror : mkfloat RMS error in fit to offset - i.e. δ cosinefiterror : mkfloat Maximum of the RMS errors of the cosine fits at each time point amplitude : Array{mkfloat, 1} Values of amplitude from which growthrate fit was calculated offset : Array{mk_float, 1} Values of offset from which frequency fit was calculated

moment_kinetics.post_processing.init_postprocessing_options — Method

moment_kinetics.post_processing.moving_average — Method

Calculate a moving average

result[i] = mean(v[i-n:i+n])

Except near the ends of the array where indices outside the range of v are skipped.

moment_kinetics.post_processing.plot_1D_1V_diagnostics — Method

moment_kinetics.post_processing.plot_fields — Method

moment_kinetics.post_processing.plot_moments — Method