time_advance

moment_kinetics.time_advance — Module

moment_kinetics.time_advance.euler_time_advance! — Method

eulertimeadvance! advances the vector equation dfvec/dt = G[f] that includes the kinetic equation + any evolved moment equations using the forward Euler method: fvecout = fvecin + dt*fvecin, with fvecin an input and fvec_out the output

moment_kinetics.time_advance.normalize_pdf! — Method

if evolving the density via continuity equation, redefine the normalised f → f/n if evolving the parallel pressure via energy equation, redefine f -> f * vth / n 'scratch' should be a (nz,nspecies) array

moment_kinetics.time_advance.rk_update! — Method

use information obtained from the Runge-Kutta stages to compute the updated pdf; for the quantities (density, upar, ppar, vth, qpar and phi) that are derived from the 'true', un-modified pdf, either: update them using info from Runge Kutta stages, if the quantities are evolved separately from the modified pdf; or update them by taking the appropriate velocity moment of the evolved pdf

moment_kinetics.time_advance.rk_update_evolved_moments! — Method

use Runge Kutta to update any charged-particle velocity moments evolved separately from the pdf

moment_kinetics.time_advance.rk_update_evolved_moments_neutral! — Method

use Runge Kutta to update any neutral-particle velocity moments evolved separately from the pdf

moment_kinetics.time_advance.setup_advance_flags — Method

create the 'advance_info' struct to be used in later Euler advance to indicate which parts of the equations are to be advanced concurrently. if no splitting of operators, all terms advanced concurrently; else, will advance one term at a time.

moment_kinetics.time_advance.setup_runge_kutta_coefficients — Method

given the number of Runge Kutta stages that are requested, returns the needed Runge Kutta coefficients; e.g., if f is the function to be updated, then f^{n+1}[stage+1] = rkcoef[1,stage]*f^{n} + rkcoef[2,stage]f^{n+1}[stage] + rk_coef[3,stage](f^{n}+dt*G[f^{n+1}[stage]]

moment_kinetics.time_advance.setup_scratch_arrays — Method

create an array of structs containing scratch arrays for the normalised pdf and low-order moments that may be evolved separately via fluid equations

moment_kinetics.time_advance.setup_time_advance! — Method

create arrays and do other work needed to setup the main time advance loop. this includes creating and populating structs for Chebyshev transforms, velocity space moments, EM fields, and advection terms

moment_kinetics.time_advance.ssp_rk! — Method

moment_kinetics.time_advance.time_advance! — Method

solve ∂f/∂t + v(z,t)⋅∂f/∂z + dvpa/dt ⋅ ∂f/∂vpa= 0 define approximate characteristic velocity v₀(z)=vⁿ(z) and take time derivative along this characteristic df/dt + δv⋅∂f/∂z = 0, with δv(z,t)=v(z,t)-v₀(z) for prudent choice of v₀, expect δv≪v so that explicit time integrator can be used without severe CFL condition

moment_kinetics.time_advance.time_advance_no_splitting! — Method

moment_kinetics.time_advance.time_advance_split_operators! — Method

moment_kinetics.time_advance.update_derived_moments! — Method

update velocity moments that are calculable from the evolved charged pdf

moment_kinetics.time_advance.update_derived_moments_neutral! — Method

update velocity moments that are calculable from the evolved neutral pdf

moment_kinetics.time_advance.update_solution_vector! — Method

update the vector containing the pdf and any evolved moments of the pdf for use in the Runge-Kutta time advance