velocity_moments

moment_kinetics.velocity_moments — Module

moment_kinetics.velocity_moments.moments — Type

moment_kinetics.velocity_moments.create_moments — Method

moment_kinetics.velocity_moments.enforce_moment_constraints! — Method

moment_kinetics.velocity_moments.integrate_over_negative_vpa — Method

computes the integral over vpa <= 0 of the integrand, using the input vpa_wgts this could be made more efficient for the case that dz/dt = vpa is time-independent, but it has been left general for the cases where, e.g., dz/dt = wpa*vth + upar varies in time

moment_kinetics.velocity_moments.integrate_over_positive_vpa — Method

computes the integral over vpa >= 0 of the integrand, using the input vpa_wgts this could be made more efficient for the case that dz/dt = vpa is time-independent, but it has been left general for the cases where, e.g., dz/dt = wpa*vth + upar varies in time

moment_kinetics.velocity_moments.integrate_over_vspace — Method

computes the integral over vpa of the integrand, using the input vpa_wgts

moment_kinetics.velocity_moments.reset_moments_status! — Method

moment_kinetics.velocity_moments.update_density! — Method

NB: if this function is called and if dens_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_density_species! — Method

calculate the updated density (dens) for a given species

moment_kinetics.velocity_moments.update_moments! — Method

calculate the updated density (dens) and parallel pressure (ppar) for all species

moment_kinetics.velocity_moments.update_ppar! — Method

NB: if this function is called and if ppar_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_ppar_species! — Method

calculate the updated parallel pressure (ppar) for a given species

moment_kinetics.velocity_moments.update_qpar! — Method

NB: if this function is called and if ppar_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_qpar_species! — Method

calculate the updated parallel heat flux (qpar) for a given species

moment_kinetics.velocity_moments.update_upar! — Method

NB: if this function is called and if upar_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_upar_species! — Method

calculate the updated parallel flow (upar) for a given species