fokker_planck

moment_kinetics.fokker_planck — Module

module for including the Full-F Fokker-Planck Collision Operator

The functions in this module are split into two groups.

The first set of functions implement the weak-form Collision operator using the Rosenbluth-MacDonald-Judd formulation in a divergence form. The Green's functions for the Rosenbluth potentials are used to obtain the Rosenbluth potentials at the boundaries. To find the potentials everywhere else elliptic solves of the PDEs for the Rosenbluth potentials are performed with Dirichlet boundary conditions. These routines provide the default collision operator used in the code.

The second set of functions are used to set up the necessary arrays to compute the Rosenbluth potentials everywhere in vpa, vperp by direct integration of the Green's functions. These functions are supported for the purposes of testing and debugging.

Lower-level routines are provided by functions from moment_kinetics.fokker_planck_calculus.

Parallelisation of the collision operator uses a special 'anyv' region type, see Collision operator and anyv region.

moment_kinetics.fokker_planck.allocate_fokkerplanck_arrays_direct_integration — Method

allocate the required ancilliary arrays

moment_kinetics.fokker_planck.explicit_fokker_planck_collisions_weak_form! — Method

Function for advancing with the explicit, weak-form, self-collision operator

moment_kinetics.fokker_planck.fokker_planck_collision_operator_weak_form! — Method

Function for evaluating $C_{ss'} = C_{ss'}[F_s,F_{s'}]$

The result is stored in the array fkpl_arrays.CC.

The normalised collision frequency is defined by

\[\tilde{\nu}_{ss'} = \frac{L_{\mathrm{ref}}}{c_{\mathrm{ref}}}\frac{\gamma_{ss'} n_\mathrm{ref}}{m_s^2 c_\mathrm{ref}^3}\]

with $\gamma_{ss'} = 2 \pi (Z_s Z_{s'})^2 e^4 \ln \Lambda_{ss'} / (4 \pi \epsilon_0)^2$.

moment_kinetics.fokker_planck.init_fokker_planck_collisions_direct_integration — Method

function that initialises the arrays needed to calculate the Rosenbluth potentials by direct integration. As this function is only supported to keep the testing of the direct integration method, the struct 'fka' created here does not contain all of the arrays necessary to compute the weak-form operator. This functionality could be ported if necessary.

moment_kinetics.fokker_planck.init_fokker_planck_collisions_weak_form — Method

function that initialises the arrays needed for Fokker Planck collisions using numerical integration to compute the Rosenbluth potentials only at the boundary and using an elliptic solve to obtain the potentials in the rest of the velocity space domain.

moment_kinetics.fokker_planck.setup_fkpl_collisions_input — Method

Function for reading Fokker Planck collision operator input parameters. Structure the namelist as follows.

[fokkerplanckcollisions] usefokkerplanck = true nuii = 1.0 frequency_option = "manual"