initial_conditions

moment_kinetics.initial_conditions — Module

moment_kinetics.initial_conditions.pdf_substruct — Type

moment_kinetics.initial_conditions.allocate_pdf_and_moments — Method

Creates the structs for the pdf and the velocity-space moments

moment_kinetics.initial_conditions.convert_full_f_charged_to_normalised! — Method

Take the full charged-particle distribution function, calculate the moments, then normalise and shift to the moment-kinetic grid.

Uses input value of f and modifies in place to the normalised distribution functions. Input density, upar, ppar, and vth are not used, the values are overwritten with the moments of f.

Inputs/outputs depend on z, vperp, and vpa (should be inside loops over species, r)

moment_kinetics.initial_conditions.convert_full_f_neutral_to_normalised! — Method

Take the full neutral-particle distribution function, calculate the moments, then normalise and shift to the moment-kinetic grid.

Uses input value of f and modifies in place to the normalised distribution functions. Input density, upar, ppar, and vth are not used, the values are overwritten with the moments of f.

Inputs/outputs depend on z, vzeta, vr and vz (should be inside loops over species, r)

moment_kinetics.initial_conditions.create_boundary_distributions — Method

Allocate arrays for distributions to be applied as boundary conditions to the pdf at various boundaries. Also initialise the Knudsen cosine distribution here so it can be used when initialising the neutral pdf.

moment_kinetics.initial_conditions.create_pdf — Method

Allocate arrays for pdfs

moment_kinetics.initial_conditions.enforce_boundary_conditions! — Method

enforce boundary conditions in vpa and z on the evolved pdf; also enforce boundary conditions in z on all separately evolved velocity space moments of the pdf

moment_kinetics.initial_conditions.enforce_initial_tapered_zero_incoming! — Method

Set up an initial condition that tries to be smoothly compatible with the sheath boundary condition for ions, by setting f(±(v_parallel-u0)<0) where u0=0 at the sheath boundaries and for z<0 increases linearly to u0=vpa.L at z=0, while for z>0 increases from u0=-vpa.L at z=0 to zero at the z=z.L/2 sheath.

To be applied to 'full-f' distribution function on vparallel grid (not wparallel grid).

moment_kinetics.initial_conditions.enforce_neutral_boundary_conditions! — Method

enforce boundary conditions on neutral particle distribution function

moment_kinetics.initial_conditions.enforce_neutral_wall_bc! — Method

enforce the wall boundary condition on neutrals; i.e., the incoming flux of neutrals equals the sum of the ion/neutral outgoing fluxes

moment_kinetics.initial_conditions.enforce_neutral_z_boundary_condition! — Method

enforce boundary conditions on neutral particle f in z

moment_kinetics.initial_conditions.enforce_r_boundary_condition! — Method

enforce boundary conditions on f in r

moment_kinetics.initial_conditions.enforce_v_boundary_condition_local! — Method

moment_kinetics.initial_conditions.enforce_vpa_boundary_condition! — Method

impose the prescribed vpa boundary condition on f at every z grid point

moment_kinetics.initial_conditions.enforce_z_boundary_condition! — Method

enforce boundary conditions on charged particle f in z

moment_kinetics.initial_conditions.enforce_z_boundary_condition_moments! — Method

enforce the z boundary condition on the evolved velocity space moments of f

moment_kinetics.initial_conditions.enforce_zero_incoming_bc! — Method

enforce a zero incoming BC in z for given species pdf at each radial location

moment_kinetics.initial_conditions.init_charged_pdf_over_density! — Method

moment_kinetics.initial_conditions.init_density! — Method

moment_kinetics.initial_conditions.init_neutral_pdf_over_density! — Method

moment_kinetics.initial_conditions.init_pdf_and_moments! — Method

creates the normalised pdf and the velocity-space moments and populates them with a self-consistent initial condition

moment_kinetics.initial_conditions.init_upar! — Method

moment_kinetics.initial_conditions.init_uz! — Method

moment_kinetics.initial_conditions.init_vth! — Method

for now the only initialisation option for the temperature is constant in z returns vth0 = sqrt(2Ts/ms) / sqrt(2Te/ms) = sqrt(Ts/Te)

moment_kinetics.initial_conditions.initialize_pdf! — Method

moment_kinetics.initial_conditions.vpagrid_to_dzdt — Method

create an array of dz/dt values corresponding to the given vpagrid values