input_structs

moment_kinetics.input_structs — Module

moment_kinetics.input_structs.advance_info — Type

moment_kinetics.input_structs.advection_input — Type

moment_kinetics.input_structs.advection_input_mutable — Type

moment_kinetics.input_structs.collisions_input — Type

moment_kinetics.input_structs.drive_input — Type

moment_kinetics.input_structs.drive_input_mutable — Type

moment_kinetics.input_structs.electron_physics_type — Type

moment_kinetics.input_structs.evolve_moments_options — Type

moment_kinetics.input_structs.geometry_input — Type

moment_kinetics.input_structs.grid_input — Type

moment_kinetics.input_structs.grid_input_mutable — Type

moment_kinetics.input_structs.initial_condition_input — Type

moment_kinetics.input_structs.initial_condition_input_mutable — Type

moment_kinetics.input_structs.io_input — Type

Settings and input for setting up file I/O

moment_kinetics.input_structs.pp_input — Type

moment_kinetics.input_structs.species_composition — Type

moment_kinetics.input_structs.species_parameters — Type

moment_kinetics.input_structs.species_parameters_mutable — Type

moment_kinetics.input_structs.time_input — Type

Base.get — Method

Utility method for converting a string to an Enum when getting from a Dict, based on the type of the default value

moment_kinetics.input_structs.Dict_to_NamedTuple — Method

Convert a Dict whose keys are String or Symbol to a NamedTuple

Useful as NamedTuple is immutable, so option values cannot be accidentally changed.

moment_kinetics.input_structs.set_defaults_and_check_section! — Method

Set the defaults for options in a section, and check that there are not any unexpected options (i.e. options that have no default).

Modifies the options[section_name]::Dict by adding defaults for any values that are not already present.

moment_kinetics.input_structs.set_defaults_and_check_top_level! — Method

Set the defaults for options in the top level of the input, and check that there are not any unexpected options (i.e. options that have no default).

Modifies the options[section_name]::Dict by adding defaults for any values that are not already present.

Ignores any sections, as these will be checked separately.