velocity_moments

moment_kinetics.velocity_moments — Module

moment_kinetics.velocity_moments.moments_charged_substruct — Type

moment_kinetics.velocity_moments.moments_neutral_substruct — Type

moment_kinetics.velocity_moments.calculate_moment_derivatives! — Method

Pre-calculate spatial derivatives of the moments that will be needed for the time advance

moment_kinetics.velocity_moments.calculate_moment_derivatives_neutral! — Method

Pre-calculate spatial derivatives of the neutral moments that will be needed for the time advance

moment_kinetics.velocity_moments.create_moments_charged — Method

moment_kinetics.velocity_moments.get_pressure — Method

compute the isotropic pressure from the already computed ppar and pperp

moment_kinetics.velocity_moments.integrate_over_negative_vpa — Method

computes the integral over vpa <= 0 of the integrand, using the input vpa_wgts this could be made more efficient for the case that dz/dt = vpa is time-independent, but it has been left general for the cases where, e.g., dz/dt = wpa*vth + upar varies in time

moment_kinetics.velocity_moments.integrate_over_positive_vpa — Method

computes the integral over vpa >= 0 of the integrand, using the input vpa_wgts this could be made more efficient for the case that dz/dt = vpa is time-independent, but it has been left general for the cases where, e.g., dz/dt = wpa*vth + upar varies in time

moment_kinetics.velocity_moments.integrate_over_vspace — Method

computes the integral over vpa of the integrand, using the input vpa_wgts

moment_kinetics.velocity_moments.reset_moments_status! — Method

moment_kinetics.velocity_moments.update_chodura_integral_species! — Method

compute the integral needed for the generalised Chodura condition

IChodura = (Z^2 vBohm^2 / cref^2) * int ( f bz^2 / vz^2 + dfdr*rhostar/vz ) vBohm = sqrt(Z Te/mi) with Z = 1 and mref = mi cref = sqrt(2Ti/mi) and normalise to the local ion density, appropriate to assessing the Chodura condition

IChodura <= (Te/e)d ne / dphi |(sheath entrance) = ni

to a single species plasma with Z = 1

moment_kinetics.velocity_moments.update_density! — Method

NB: if this function is called and if dens_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_density_species! — Method

calculate the updated density (dens) for a given species; should only be called when evolvedensity = false, in which case the vpa coordinate is vpa/cs

moment_kinetics.velocity_moments.update_moments! — Method

calculate the updated density (dens) and parallel pressure (ppar) for all species

moment_kinetics.velocity_moments.update_moments_neutral! — Method

update velocity moments of the evolved neutral pdf

moment_kinetics.velocity_moments.update_neutral_density! — Method

calculate the neutral density from the neutral pdf

moment_kinetics.velocity_moments.update_neutral_density_species! — Method

calculate the updated density (dens) for a given species

moment_kinetics.velocity_moments.update_neutral_pr_species! — Method

calculate the updated pressure in the rr direction (pr) for a given species

moment_kinetics.velocity_moments.update_neutral_pz_species! — Method

calculate the updated pressure in zz direction (pz) for a given species

moment_kinetics.velocity_moments.update_neutral_pzeta_species! — Method

calculate the updated pressure in the zeta zeta direction (pzeta) for a given species

moment_kinetics.velocity_moments.update_neutral_qz_species! — Method

calculate the updated heat flux zzz direction (qz) for a given species

moment_kinetics.velocity_moments.update_neutral_ur_species! — Method

calculate the updated ur (mean velocity in r) for a given species

moment_kinetics.velocity_moments.update_neutral_uz_species! — Method

calculate the updated uz (mean velocity in z) for a given species

moment_kinetics.velocity_moments.update_neutral_uzeta_species! — Method

calculate the updated uzeta (mean velocity in zeta) for a given species

moment_kinetics.velocity_moments.update_ppar! — Method

NB: if this function is called and if ppar_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_ppar_species! — Method

calculate the updated energy density (or parallel pressure, ppar) for a given species; which of these is calculated depends on the definition of the vpa coordinate

moment_kinetics.velocity_moments.update_pperp_species! — Method

calculate the updated perpendicular pressure (pperp) for a given species

moment_kinetics.velocity_moments.update_qpar! — Method

NB: the incoming pdf is the normalized pdf

moment_kinetics.velocity_moments.update_qpar_species! — Method

calculate the updated parallel heat flux (qpar) for a given species

moment_kinetics.velocity_moments.update_upar! — Method

NB: if this function is called and if upar_updated is false, then the incoming pdf is the un-normalized pdf that satisfies int dv pdf = density

moment_kinetics.velocity_moments.update_upar_species! — Method

calculate the updated parallel flow (upar) for a given species